requirements_bak.txt
matplotlib==3.4.2
numpy==1.19.5
starfile==0.4.5
tensorflow==2.5.0
scipy==1.7.0
mrcfile==1.3.0
ipython==7.28.0
Pillow==8.4.0
